Create New Programmer Participation
This is a file to help the owner of the repository think more inclusively toward new programmers as he charts his path toward a Hackathon of his own. clarity as to what a contribution will help all programmers, but it will immensely help new programmers who may feel excluded if the right wording isn't employed in advertisement, or listed clearly on the repository site. New coders offer a unique perspective, they are the future technologists.